requirements.txt
asn1crypto==0.24.0
awsebcli==3.15.3
backports.functools-lru-cache==1.6.1
backports.ssl-match-hostname==3.7.0.1
beautifulsoup4==4.8.2
blessed==1.15.0
botocore==1.12.239
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
cryptography==2.1.4
defusedxml==0.4.1
Django==1.11.28
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
idna==2.7
ipaddress==1.0.22
jmespath==0.9.4
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
oauthlib==3.1.0
pathspec==0.5.9
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.3.0
SecretStorage==2.3.1
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
